syrconnect2mqtt/syrconnect2mqtt/requirements.txt
paho-mqtt
requests
xmltodict
pycryptodome
